fix graph colours, restructure the files, add routes for pages
