+++ src/component/chart/DoughnutChart.jsx
@@ -1,18 +1,22 @@
 import React from "react";
 import { Doughnut } from 'react-chartjs-2';
+import Chart from 'chart.js/auto';
+import { CategoryScale } from "chart.js";
+import { Colors } from 'chart.js';
+
+Chart.register(CategoryScale);
+Chart.register(Colors);
 
 const DoughnutChart = ({ chartData }) => {
 
     return (
       <div className="chart-container">
-        <h2 style={{ textAlign: "center", color: 'black' }}>Expense Chart</h2>
-        {/* <span>{chartData.dataset[0].data}</span> */}
         <Doughnut
           data={ chartData }
           options={{
             plugins: {
               colors: {
-                enabled: true
+                forceOverride: true
               }
             }
           }}
