+++ src/utils/Data.jsx
@@ -1,27 +1,3 @@
 export const Expense = [
-    {
-      category: 'Rent',
-      amount: 1500,
-      date: '',
-    },
-    {
-      category: 'Food',
-      amount: 500,
-      date: '',
-    },
-    {
-      category: 'Leisure',
-      amount: 100,
-      date: '',
-    },
-    {
-      category: 'Taxes',
-      amount: 250,
-      date: '',
-    },
-    {
-      category: 'Utilities',
-      amount: 190,
-      date: '',
-    },
+    
   ]
